fix(CAT-43): header and pages layout
