+++ packages/client/src/components/error-container/error-container.css
@@ -1,7 +1,3 @@
-.error-container {
-  padding-top: 14%;
-}
-
 .error-container__title {
   margin: 0;
   font-size: 96px;
