+++ packages/client/src/components/exit-button/exit-button.css
@@ -2,6 +2,7 @@
   display: flex;
   justify-content: flex-end;
   align-self: flex-end;
+  padding-right: 30px;
 }
 
 .exit-button__button {
