+++ packages/client/src/components/header/header.css
@@ -1,0 +1,7 @@
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 30px;
+  margin-bottom: 30px;
+}
