+++ packages/client/src/pages/profile-page/profile-page.tsx
@@ -12,24 +12,26 @@ const ProfilePage = () => {
 
   return (
     <PageWrapper>
-      <div
-        className={[
-          'profile-page',
-          theme === Theme.Dark ? 'profile-page--dark' : null,
-        ].join(' ')}>
-        <div className='profile-page__container'>
-          <ProfileAvatar />
-          <ProfileForm />
-          <p
-            className='profile-page__change-password'
-            onClick={() => navigate('../change-password')}>
-            Сменить пароль
-          </p>
+      <>
+        <div
+          className={[
+            'profile-page',
+            theme === Theme.Dark ? 'profile-page--dark' : null,
+          ].join(' ')}>
+          <div className='profile-page__container'>
+            <ProfileAvatar />
+            <ProfileForm />
+            <p
+              className='profile-page__change-password'
+              onClick={() => navigate('../change-password')}>
+              Сменить пароль
+            </p>
+          </div>
         </div>
         <div className='profile-page__exit'>
           <ExitButton />
         </div>
-      </div>
+      </>
     </PageWrapper>
   );
 };
